opt/apps/server-a/code/image-backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
pymongo>=4.6,<5.0
google-auth>=2.29,<3.0
requests>=2.31,<3.0
boto3>=1.34,<2.0
Pillow>=10.2,<11.0
gunicorn>=21.2,<22.0
